+++ tests/Amadeus/Client/Struct/Service/IntegratedPricingTest.php
@@ -22,6 +22,8 @@
 
 namespace Test\Amadeus\Client\Struct\Service;
 
+use Amadeus\Client\RequestOptions\Service\FormOfPayment;
+use Amadeus\Client\RequestOptions\Service\FrequentFlyer;
 use Amadeus\Client\RequestOptions\Service\PaxSegRef;
 use Amadeus\Client\RequestOptions\ServiceIntegratedPricingOptions;
 use Amadeus\Client\Struct\Fare\PricePnr13\DateInformation;
@@ -43,20 +45,20 @@ class IntegratedPricingTest extends BaseTestCase
     {
         $opt = new ServiceIntegratedPricingOptions();
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_NO_OPTION, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_NO_OPTION, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertNull($message->pricingOption[0]->optionDetail);
-        $this->assertNull($message->pricingOption[0]->carrierInformation);
-        $this->assertNull($message->pricingOption[0]->currency);
-        $this->assertNull($message->pricingOption[0]->dateInformation);
-        $this->assertNull($message->pricingOption[0]->formOfPaymentInformation);
-        $this->assertNull($message->pricingOption[0]->frequentFlyerInformation);
-        $this->assertNull($message->pricingOption[0]->locationInformation);
-        $this->assertNull($message->pricingOption[0]->paxSegTstReference);
-        $this->assertNull($message->pricingOption[0]->ticketInformation);
+        $this->assertNull($msg->pricingOption[0]->optionDetail);
+        $this->assertNull($msg->pricingOption[0]->carrierInformation);
+        $this->assertNull($msg->pricingOption[0]->currency);
+        $this->assertNull($msg->pricingOption[0]->dateInformation);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation);
+        $this->assertNull($msg->pricingOption[0]->frequentFlyerInformation);
+        $this->assertNull($msg->pricingOption[0]->locationInformation);
+        $this->assertNull($msg->pricingOption[0]->paxSegTstReference);
+        $this->assertNull($msg->pricingOption[0]->ticketInformation);
     }
 
     public function testCanMakeMessageWithAccountCode()
@@ -71,26 +73,26 @@ class IntegratedPricingTest extends BaseTestCase
             ]
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
+        $this->assertCount(1, $msg->pricingOption);
 
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_ACCOUNT_CODE, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
-        $this->assertCount(1, $message->pricingOption[0]->optionDetail->criteriaDetails);
-        $this->assertEquals('AAA123456', $message->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
-        $this->assertNull($message->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeDescription);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_ACCOUNT_CODE, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption[0]->optionDetail->criteriaDetails);
+        $this->assertEquals('AAA123456', $msg->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
+        $this->assertNull($msg->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeDescription);
 
-        $this->assertCount(1, $message->pricingOption[0]->paxSegTstReference->referenceDetails);
-        $this->assertEquals(1, $message->pricingOption[0]->paxSegTstReference->referenceDetails[0]->value);
-        $this->assertEquals(ReferenceDetails::QUALIFIER_PAX_REFERENCE, $message->pricingOption[0]->paxSegTstReference->referenceDetails[0]->type);
+        $this->assertCount(1, $msg->pricingOption[0]->paxSegTstReference->referenceDetails);
+        $this->assertEquals(1, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[0]->value);
+        $this->assertEquals(ReferenceDetails::QUALIFIER_PAX_REFERENCE, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[0]->type);
 
-        $this->assertNull($message->pricingOption[0]->carrierInformation);
-        $this->assertNull($message->pricingOption[0]->currency);
-        $this->assertNull($message->pricingOption[0]->dateInformation);
-        $this->assertNull($message->pricingOption[0]->formOfPaymentInformation);
-        $this->assertNull($message->pricingOption[0]->frequentFlyerInformation);
-        $this->assertNull($message->pricingOption[0]->locationInformation);
-        $this->assertNull($message->pricingOption[0]->ticketInformation);
+        $this->assertNull($msg->pricingOption[0]->carrierInformation);
+        $this->assertNull($msg->pricingOption[0]->currency);
+        $this->assertNull($msg->pricingOption[0]->dateInformation);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation);
+        $this->assertNull($msg->pricingOption[0]->frequentFlyerInformation);
+        $this->assertNull($msg->pricingOption[0]->locationInformation);
+        $this->assertNull($msg->pricingOption[0]->ticketInformation);
     }
 
     /**
@@ -115,30 +117,30 @@ class IntegratedPricingTest extends BaseTestCase
             ]
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
+        $this->assertCount(1, $msg->pricingOption);
 
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_PAX_SEG_ELEMENT_SELECTION, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_PAX_SEG_ELEMENT_SELECTION, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertNull($message->pricingOption[0]->optionDetail);
+        $this->assertNull($msg->pricingOption[0]->optionDetail);
 
-        $this->assertCount(3, $message->pricingOption[0]->paxSegTstReference->referenceDetails);
-        $this->assertEquals(1, $message->pricingOption[0]->paxSegTstReference->referenceDetails[0]->value);
-        $this->assertEquals(ReferenceDetails::QUALIFIER_PAX_REFERENCE, $message->pricingOption[0]->paxSegTstReference->referenceDetails[0]->type);
-        $this->assertEquals(2, $message->pricingOption[0]->paxSegTstReference->referenceDetails[1]->value);
-        $this->assertEquals(ReferenceDetails::QUALIFIER_SEGMENT_REFERENCE, $message->pricingOption[0]->paxSegTstReference->referenceDetails[1]->type);
-        $this->assertEquals(16, $message->pricingOption[0]->paxSegTstReference->referenceDetails[2]->value);
-        $this->assertEquals(ReferenceDetails::QUALIFIER_ELEMENT, $message->pricingOption[0]->paxSegTstReference->referenceDetails[2]->type);
+        $this->assertCount(3, $msg->pricingOption[0]->paxSegTstReference->referenceDetails);
+        $this->assertEquals(1, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[0]->value);
+        $this->assertEquals(ReferenceDetails::QUALIFIER_PAX_REFERENCE, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[0]->type);
+        $this->assertEquals(2, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[1]->value);
+        $this->assertEquals(ReferenceDetails::QUALIFIER_SEGMENT_REFERENCE, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[1]->type);
+        $this->assertEquals(16, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[2]->value);
+        $this->assertEquals(ReferenceDetails::QUALIFIER_ELEMENT, $msg->pricingOption[0]->paxSegTstReference->referenceDetails[2]->type);
 
-        $this->assertNull($message->pricingOption[0]->optionDetail);
-        $this->assertNull($message->pricingOption[0]->carrierInformation);
-        $this->assertNull($message->pricingOption[0]->currency);
-        $this->assertNull($message->pricingOption[0]->dateInformation);
-        $this->assertNull($message->pricingOption[0]->formOfPaymentInformation);
-        $this->assertNull($message->pricingOption[0]->frequentFlyerInformation);
-        $this->assertNull($message->pricingOption[0]->locationInformation);
-        $this->assertNull($message->pricingOption[0]->ticketInformation);
+        $this->assertNull($msg->pricingOption[0]->optionDetail);
+        $this->assertNull($msg->pricingOption[0]->carrierInformation);
+        $this->assertNull($msg->pricingOption[0]->currency);
+        $this->assertNull($msg->pricingOption[0]->dateInformation);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation);
+        $this->assertNull($msg->pricingOption[0]->frequentFlyerInformation);
+        $this->assertNull($msg->pricingOption[0]->locationInformation);
+        $this->assertNull($msg->pricingOption[0]->ticketInformation);
     }
 
     public function testCanMakeMessageWithOverrideOptionServicePricingDescription()
@@ -149,21 +151,21 @@ class IntegratedPricingTest extends BaseTestCase
             ]
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
+        $this->assertCount(1, $msg->pricingOption);
 
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_SHOW_PRICING_DESCRIPTION, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_SHOW_PRICING_DESCRIPTION, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertNull($message->pricingOption[0]->optionDetail);
-        $this->assertNull($message->pricingOption[0]->carrierInformation);
-        $this->assertNull($message->pricingOption[0]->currency);
-        $this->assertNull($message->pricingOption[0]->dateInformation);
-        $this->assertNull($message->pricingOption[0]->formOfPaymentInformation);
-        $this->assertNull($message->pricingOption[0]->frequentFlyerInformation);
-        $this->assertNull($message->pricingOption[0]->locationInformation);
-        $this->assertNull($message->pricingOption[0]->paxSegTstReference);
-        $this->assertNull($message->pricingOption[0]->ticketInformation);
+        $this->assertNull($msg->pricingOption[0]->optionDetail);
+        $this->assertNull($msg->pricingOption[0]->carrierInformation);
+        $this->assertNull($msg->pricingOption[0]->currency);
+        $this->assertNull($msg->pricingOption[0]->dateInformation);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation);
+        $this->assertNull($msg->pricingOption[0]->frequentFlyerInformation);
+        $this->assertNull($msg->pricingOption[0]->locationInformation);
+        $this->assertNull($msg->pricingOption[0]->paxSegTstReference);
+        $this->assertNull($msg->pricingOption[0]->ticketInformation);
     }
 
     public function testCanMakeMessageWithOverridePointOfSale()
@@ -172,24 +174,24 @@ class IntegratedPricingTest extends BaseTestCase
             'pointOfSaleOverride' => 'MUC'
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
+        $this->assertCount(1, $msg->pricingOption);
 
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_POINT_OF_SALE, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_POINT_OF_SALE, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertEquals(LocationInformation::TYPE_POINT_OF_SALE, $message->pricingOption[0]->locationInformation->locationType);
-        $this->assertEquals('MUC', $message->pricingOption[0]->locationInformation->firstLocationDetails->code);
-        $this->assertNull($message->pricingOption[0]->locationInformation->secondLocationDetails);
+        $this->assertEquals(LocationInformation::TYPE_POINT_OF_SALE, $msg->pricingOption[0]->locationInformation->locationType);
+        $this->assertEquals('MUC', $msg->pricingOption[0]->locationInformation->firstLocationDetails->code);
+        $this->assertNull($msg->pricingOption[0]->locationInformation->secondLocationDetails);
 
-        $this->assertNull($message->pricingOption[0]->optionDetail);
-        $this->assertNull($message->pricingOption[0]->carrierInformation);
-        $this->assertNull($message->pricingOption[0]->currency);
-        $this->assertNull($message->pricingOption[0]->dateInformation);
-        $this->assertNull($message->pricingOption[0]->formOfPaymentInformation);
-        $this->assertNull($message->pricingOption[0]->frequentFlyerInformation);
-        $this->assertNull($message->pricingOption[0]->paxSegTstReference);
-        $this->assertNull($message->pricingOption[0]->ticketInformation);
+        $this->assertNull($msg->pricingOption[0]->optionDetail);
+        $this->assertNull($msg->pricingOption[0]->carrierInformation);
+        $this->assertNull($msg->pricingOption[0]->currency);
+        $this->assertNull($msg->pricingOption[0]->dateInformation);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation);
+        $this->assertNull($msg->pricingOption[0]->frequentFlyerInformation);
+        $this->assertNull($msg->pricingOption[0]->paxSegTstReference);
+        $this->assertNull($msg->pricingOption[0]->ticketInformation);
     }
 
     public function testCanMakeMessageWithPricingDateOverride()
@@ -202,14 +204,14 @@ class IntegratedPricingTest extends BaseTestCase
             )
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_PRICING_DATE, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
-        $this->assertEquals(DateInformation::OPT_DATE_OVERRIDE, $message->pricingOption[0]->dateInformation->businessSemantic);
-        $this->assertEquals("2012", $message->pricingOption[0]->dateInformation->dateTime->year);
-        $this->assertEquals("06", $message->pricingOption[0]->dateInformation->dateTime->month);
-        $this->assertEquals("27", $message->pricingOption[0]->dateInformation->dateTime->day);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_PRICING_DATE, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertEquals(DateInformation::OPT_DATE_OVERRIDE, $msg->pricingOption[0]->dateInformation->businessSemantic);
+        $this->assertEquals("2012", $msg->pricingOption[0]->dateInformation->dateTime->year);
+        $this->assertEquals("06", $msg->pricingOption[0]->dateInformation->dateTime->month);
+        $this->assertEquals("27", $msg->pricingOption[0]->dateInformation->dateTime->day);
     }
 
     public function testCanMakeMessageWithCurrencyOverride()
@@ -218,13 +220,13 @@ class IntegratedPricingTest extends BaseTestCase
             'currencyOverride' => 'EUR'
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_CURRENCY, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_CURRENCY, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertEquals(FirstCurrencyDetails::QUAL_CURRENCY_OVERRIDE, $message->pricingOption[0]->currency->firstCurrencyDetails->currencyQualifier);
-        $this->assertEquals('EUR', $message->pricingOption[0]->currency->firstCurrencyDetails->currencyIsoCode);
+        $this->assertEquals(FirstCurrencyDetails::QUAL_CURRENCY_OVERRIDE, $msg->pricingOption[0]->currency->firstCurrencyDetails->currencyQualifier);
+        $this->assertEquals('EUR', $msg->pricingOption[0]->currency->firstCurrencyDetails->currencyIsoCode);
     }
 
     public function testCanMakeMessageWithValidatingCarrierOverride()
@@ -233,12 +235,12 @@ class IntegratedPricingTest extends BaseTestCase
             'validatingCarrier' => 'BA'
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_VALIDATING_CARRIER, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_VALIDATING_CARRIER, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertEquals('BA', $message->pricingOption[0]->carrierInformation->companyIdentification->otherCompany);
+        $this->assertEquals('BA', $msg->pricingOption[0]->carrierInformation->companyIdentification->otherCompany);
     }
 
     public function testCanMakeMessageWithAwardPricing()
@@ -247,12 +249,12 @@ class IntegratedPricingTest extends BaseTestCase
             'awardPricing' => ServiceIntegratedPricingOptions::AWARDPRICING_MILES
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_AWARD, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_AWARD, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertEquals('MIL', $message->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
+        $this->assertEquals('MIL', $msg->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
     }
 
     public function testCanMakeMessageWithCorporationNumber()
@@ -261,12 +263,12 @@ class IntegratedPricingTest extends BaseTestCase
             'corporationNumber' => '123435'
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_CORPORATION_NUMBER, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_CORPORATION_NUMBER, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertEquals('123435', $message->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
+        $this->assertEquals('123435', $msg->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
     }
 
     public function testCanMakeMessageWithTicketDesignator()
@@ -275,12 +277,78 @@ class IntegratedPricingTest extends BaseTestCase
             'ticketDesignator' => '3M4J2KM432'
         ]);
 
-        $message = new IntegratedPricing($opt);
+        $msg = new IntegratedPricing($opt);
 
-        $this->assertCount(1, $message->pricingOption);
-        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_TICKET_DESIGNATOR, $message->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_TICKET_DESIGNATOR, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
 
-        $this->assertEquals('3M4J2KM432', $message->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
+        $this->assertEquals('3M4J2KM432', $msg->pricingOption[0]->optionDetail->criteriaDetails[0]->attributeType);
+    }
+
+    public function testCanMakeMessageWithFormOfPaymentOverride()
+    {
+        $opt = new ServiceIntegratedPricingOptions([
+            'formOfPayment' => [
+                new FormOfPayment([
+                    'type' => FormOfPayment::TYPE_CREDIT_CARD,
+                    'amount' => 10,
+                    'creditCardNumber' => '400000'
+                ]),
+                new FormOfPayment([
+                    'type' => FormOfPayment::TYPE_CASH
+                ]),
+            ]
+        ]);
+
+        $msg = new IntegratedPricing($opt);
+
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_FORM_OF_PAYMENT, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertEquals(\Amadeus\Client\Struct\Fare\PricePnr13\FormOfPayment::TYPE_CREDIT_CARD, $msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->type);
+        $this->assertEquals(10, $msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->amount);
+        $this->assertEquals('400000', $msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->creditCardNumber);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->sourceOfApproval);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->authorisedAmount);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->indicator);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->addressVerification);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->approvalCode);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->customerAccount);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->expiryDate);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->extendedPayment);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->fopFreeText);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->membershipStatus);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->pinCode);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->pinCodeType);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->transactionInfo);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->formOfPayment->vendorCode);
+        $this->assertCount(1, $msg->pricingOption[0]->formOfPaymentInformation->otherFormOfPayment);
+        $this->assertEquals(\Amadeus\Client\Struct\Fare\PricePnr13\FormOfPayment::TYPE_CASH, $msg->pricingOption[0]->formOfPaymentInformation->otherFormOfPayment[0]->type);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->otherFormOfPayment[0]->amount);
+        $this->assertNull($msg->pricingOption[0]->formOfPaymentInformation->otherFormOfPayment[0]->creditCardNumber);
+    }
+
+    public function testCanMakeMessageWithFrequentFlyerOverride()
+    {
+        $opt = new ServiceIntegratedPricingOptions([
+            'frequentFlyers' => [
+                new FrequentFlyer([
+                    'company' => '6X',
+                    'number' => '1234567891011',
+                    'tierLevel' => 'SILVER',
+                    'priorityCode' => '1'
+                ])
+            ]
+        ]);
+
+        $msg = new IntegratedPricing($opt);
+        $this->assertCount(1, $msg->pricingOption);
+        $this->assertEquals(IntegratedPricing\PricingOptionKey::OVERRIDE_FREQUENT_FLYER_INFORMATION, $msg->pricingOption[0]->pricingOptionKey->pricingOptionKey);
+        $this->assertCount(1, $msg->pricingOption[0]->frequentFlyerInformation->frequentTravellerDetails);
+        $this->assertEquals('6X', $msg->pricingOption[0]->frequentFlyerInformation->frequentTravellerDetails[0]->carrier);
+        $this->assertEquals('1234567891011', $msg->pricingOption[0]->frequentFlyerInformation->frequentTravellerDetails[0]->number);
+        $this->assertEquals('SILVER', $msg->pricingOption[0]->frequentFlyerInformation->frequentTravellerDetails[0]->tierLevel);
+        $this->assertEquals('1', $msg->pricingOption[0]->frequentFlyerInformation->frequentTravellerDetails[0]->priorityCode);
+
     }
 }