+++ tests/Amadeus/Client/Struct/Ticket/CreateTSMFromPricingTest.php
@@ -25,7 +25,7 @@ class CreateTSMFromPricingTest extends BaseTestCase
 {
     public function testCanMakeTstFromPricingWithPassengerReference()
     {
-        $message = new CreateTSMFromPricing(
+        $msg = new CreateTSMFromPricing(
             new TicketCreateTsmFromPricingOptions([
                 'pricings' => [
                     new Pricing([
@@ -41,20 +41,20 @@ class CreateTSMFromPricingTest extends BaseTestCase
             ])
         );
 
-        $this->assertEquals(1, count($message->psaList));
-        $this->assertEquals(2, $message->psaList[0]->itemReference->uniqueReference);
-        $this->assertEquals(ItemReference::REFTYPE_TSM, $message->psaList[0]->itemReference->referenceType);
+        $this->assertEquals(1, count($msg->psaList));
+        $this->assertEquals(2, $msg->psaList[0]->itemReference->uniqueReference);
+        $this->assertEquals(ItemReference::REFTYPE_TSM, $msg->psaList[0]->itemReference->referenceType);
 
-        $this->assertEquals(1, count($message->psaList[0]->paxReference->refDetails));
-        $this->assertEquals(1, $message->psaList[0]->paxReference->refDetails[0]->refNumber);
-        $this->assertEquals(RefDetails::QUAL_PASSENGER, $message->psaList[0]->paxReference->refDetails[0]->refQualifier);
+        $this->assertEquals(1, count($msg->psaList[0]->paxReference->refDetails));
+        $this->assertEquals(1, $msg->psaList[0]->paxReference->refDetails[0]->refNumber);
+        $this->assertEquals(RefDetails::QUAL_PASSENGER, $msg->psaList[0]->paxReference->refDetails[0]->refQualifier);
     }
 
     public function testCanMakeTsmFromPricingWithPnrInfo()
     {
         // !!! This PNR record locator is used for tracing purpose, no internal retrieve. !!!
 
-        $message = new CreateTSMFromPricing(
+        $msg = new CreateTSMFromPricing(
             new TicketCreateTsMFromPricingOptions([
                 'informationalRecordLocator' => 'ABC123',
                 'pricings' => [
@@ -65,6 +65,6 @@ class CreateTSMFromPricingTest extends BaseTestCase
             ])
         );
 
-        $this->assertEquals('ABC123', $message->pnrLocatorData->reservationInformation->controlNumber);
+        $this->assertEquals('ABC123', $msg->pnrLocatorData->reservationInformation->controlNumber);
     }
 }